Remove unserializable properties in the DialogFragmentArguments class.
To configure a WebView inside a DialogFragment and make it as flexible as possible without copying and pasting every function that WebView has, this pull request makes DialogFragment rely on Configurator Fragments which implement the WebViewDownloadConfigurator and WebViewConfigurator interfaces.

I chose to use fragments to configure the inner WebView because configurator fragments will be re-instantiated after a configuration change, and Android developers are accustomed to avoiding memory leaks.